Progress: TextAnalysisRun properties; a mock-up for results
These changes add properties to the TextAnalysisRun control for setting the various data results. The changes also include a static mock-up of results.

I want the results to show three rows of run analysis with headings on the left for each row. The headings should be fixed in place, while the results on the right will need to be scrollable (horizontally). I haven't yet got an implementation for the scrolling; everything is positioned in a 2x3 grid with horizontal stack panels for results in each row.