XamlCxPrototype1: scrolling for results
The stackpanels for dynamic results are now contained in a ScrollViewer.

The overall results display is in a 2x1 grid, with a left column for row headings and right column for the dynamically-generated results. The right column has a ScrollViewer with a vertical StackPanel for three rows of info, each row being a horizontal StackPanel. The row headings on the left are vertically aligned to the rows using static pixel positioning.