XamlCxPrototype1: replace stackpanels for results with canvas
Three rows of result tile need to be horizontally aligned according to the text starting position for each tile. That requires a canvas, so that each tile can be positioned as required. These changes replace the stackpanels used for an earlier prototype with a canvas, plus logic for calculating metrics for each result tile.