XamlCxPrototype1: fix bug; UI tweak
The tiles for bidi and merged runs were displaying like script run tiles. This was because the wrong content type was being set on those tiles. (Copilot suggestion recognized the correct pattern except for that detail and it was missed when reviewing copilot's suggestion.)

Also, a small UI tweak to reduce the padding at the bottom of the results canvas.